requirements.txt
# YouTube Shorts Automation Tool Dependencies

# Core
requests>=2.31.0
python-dotenv>=1.0.0

# YouTube API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# OpenAI for content generation
openai>=1.3.0

# Database
aiosqlite>=0.19.0

# Image processing
Pillow>=10.0.0

# HTTP for webhooks
aiohttp>=3.9.0

# Date/time utilities
pytz>=2023.3

# Video processing (optional - for local testing)
# moviepy>=1.0.3
